fix(workbench): #SKFP-103 move workbench to userdashboard
